+++ src/main/java/com/example/demo/oauth/controller/AuthController.java
@@ -1,35 +1,27 @@
-package com.example.demo.siteuser.controller;
+package com.example.demo.oauth.controller;
 
 import static com.example.demo.exception.type.ErrorCode.INVALID_NICKNAME;
-import static com.example.demo.exception.type.ErrorCode.REFRESH_TOKEN_EXPIRED;
 
 import com.example.demo.common.ResponseDto;
 import com.example.demo.common.ResponseUtil;
 import com.example.demo.exception.RacketPuncherException;
-import com.example.demo.oauth2.dto.AccessToken;
-import com.example.demo.oauth2.dto.ProfileDto;
-import com.example.demo.oauth2.service.ProviderService;
-import com.example.demo.siteuser.dto.EmailRequestDto;
-import com.example.demo.siteuser.dto.LoginResponseDto;
-import com.example.demo.siteuser.dto.NicknameRequestDto;
-import com.example.demo.siteuser.dto.QuitDto;
-import com.example.demo.siteuser.dto.AccessTokenDto;
-import com.example.demo.siteuser.dto.SignInDto;
-import com.example.demo.siteuser.dto.SignKakao;
-;
-import com.example.demo.siteuser.dto.SignUpDto;
+import com.example.demo.oauth.dto.AccessTokenDto;
+import com.example.demo.oauth.dto.EmailRequestDto;
+import com.example.demo.oauth.dto.LoginResponseDto;
+import com.example.demo.oauth.dto.NicknameRequestDto;
+import com.example.demo.oauth.dto.QuitDto;
+import com.example.demo.oauth.dto.SignInDto;
+import com.example.demo.oauth.dto.SignUpDto;
+import com.example.demo.oauth.security.TokenProvider;
+import com.example.demo.oauth.service.AuthService;
 import com.example.demo.siteuser.repository.SiteUserRepository;
-import com.example.demo.siteuser.security.TokenProvider;
-import com.example.demo.siteuser.service.MemberService;
 import java.util.concurrent.TimeUnit;
-import javax.naming.CommunicationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
-import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -37,26 +29,27 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+;
+
 @Slf4j
 @RestController
 @RequestMapping("api/auth")
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final SiteUserRepository siteUserRepository;
     private final RedisTemplate<String, String> redisTemplate;
-    private final MemberService memberService;
+    private final AuthService authService;
     private final TokenProvider tokenProvider;
-    private final ProviderService providerService;
+//    private final ProviderService providerService;
 
     @PostMapping("/sign-up")
     public void signUp(@RequestBody SignUpDto signUpDto) {
-            memberService.register(signUpDto);
+            authService.register(signUpDto);
     }
 
     @PostMapping("/sign-in")
     public ResponseDto<LoginResponseDto> signIn(@RequestBody SignInDto signInDto) {
-        memberService.authenticate(signInDto);
+        authService.authenticate(signInDto);
         var token = tokenProvider.generateAccessToken(signInDto.getEmail());
         var refreshToken = tokenProvider.generateAndSaveRefreshToken(signInDto.getEmail());
         var result = LoginResponseDto.builder()
@@ -66,43 +59,35 @@ public class AuthController {
         return ResponseUtil.SUCCESS(result);
     }
 
-    @PostMapping("/sign-in/kakao")
-    public ResponseEntity<?> signInKakao(@RequestBody SignKakao request) {
-        // 카카오용 API
-
-        try {
-            AccessToken accessToken = providerService.getAccessToken(request.getCode(), request.getProvider());
-            ProfileDto profile = providerService.getProfile(accessToken.getAccess_token(), request.getProvider());
-            var member = siteUserRepository.findByNickname(profile.getNickname());
-            if (member.isPresent()) {
-                var refreshToken = this.tokenProvider.generateAndSaveRefreshToken(member.get().getEmail());
-                var acsToken = this.tokenProvider.generateAccessToken(member.get().getEmail());
-                LoginResponseDto loginResponseDto = LoginResponseDto.builder()
-                        .accessToken(acsToken)
-                        .refreshToken(refreshToken)
-                        .build();
-                return ResponseEntity.ok(loginResponseDto);
-            } else {
-                LoginResponseDto loginResponseDto = new LoginResponseDto();
-                return ResponseEntity.ok(loginResponseDto);
-            }
-        } catch (CommunicationException e) {
-            return new ResponseEntity<>("Wrong Request", HttpStatus.BAD_REQUEST);
-        }
-    }
+//    @PostMapping("/sign-in/kakao")
+//    public ResponseEntity<?> signInKakao(@RequestBody SignKakao request) {
+//        // 카카오용 API
+//
+//        try {
+////            AccessToken accessToken = providerService.getAccessToken(request.getCode(), request.getProvider());
+////            ProfileDto profile = providerService.getProfile(accessToken.getAccess_token(), request.getProvider());
+////            var member = siteUserRepository.findByNickname(profile.getNickname());
+//            if (member.isPresent()) {
+//                var refreshToken = this.tokenProvider.generateAndSaveRefreshToken(member.get().getEmail());
+//                var acsToken = this.tokenProvider.generateAccessToken(member.get().getEmail());
+//                LoginResponseDto loginResponseDto = LoginResponseDto.builder()
+//                        .accessToken(acsToken)
+//                        .refreshToken(refreshToken)
+//                        .build();
+//                return ResponseEntity.ok(loginResponseDto);
+//            } else {
+//                LoginResponseDto loginResponseDto = new LoginResponseDto();
+//                return ResponseEntity.ok(loginResponseDto);
+//            }
+//        } catch (CommunicationException e) {
+//            return new ResponseEntity<>("Wrong Request", HttpStatus.BAD_REQUEST);
+//        }
+//    }
 
     @PostMapping("/reissue")
     public ResponseDto<AccessTokenDto> reissue(@RequestBody AccessTokenDto accessTokenDto) {
-        Authentication authentication = tokenProvider.getAuthentication(accessTokenDto.getAccessToken());
-        String refreshToken = redisTemplate.opsForValue().get(authentication.getName());
-        if (ObjectUtils.isEmpty(refreshToken)) {
-            throw new RacketPuncherException(REFRESH_TOKEN_EXPIRED);
-        }
-        var newAccessToken = tokenProvider.generateAccessToken(authentication.getName());
-        redisTemplate.delete(authentication.getName());
-        tokenProvider.generateAndSaveRefreshToken(authentication.getName());
-
-        return ResponseUtil.SUCCESS(new AccessTokenDto(newAccessToken));
+        var newAccessToken = authService.getNewAccessTokenAndSaveNewRefreshToken(accessTokenDto);
+        return ResponseUtil.SUCCESS(newAccessToken);
     }
 
     @PostMapping("/sign-out")
@@ -134,13 +119,13 @@ public class AuthController {
         Long expiration = tokenProvider.getExpiration(accessToken);
         redisTemplate.opsForValue().set(accessToken, "quit", expiration, TimeUnit.MILLISECONDS);
 
-        var result = this.memberService.withdraw(request);
+        var result = this.authService.withdraw(request);
         return new ResponseEntity<>("Quit Completed", HttpStatus.OK);
     }
 
     @PostMapping(path = "/check-email")
     public ResponseEntity<ResponseDto<String>> checkEmailExistence(@RequestBody EmailRequestDto emailRequestDto) {
-        boolean exists = memberService.isEmailExist(emailRequestDto.getEmail());
+        boolean exists = authService.isEmailExist(emailRequestDto.getEmail());
         String message = exists ? "사용 불가능한 이메일 입니다." : "사용 가능한 이메일 입니다.";
         return ResponseEntity.ok(ResponseUtil.SUCCESS(message));
     }
@@ -148,7 +133,7 @@ public class AuthController {
     @PostMapping(path = "/check-nickname")
     public ResponseEntity<ResponseDto<String>> checkNicknameExistence(
             @RequestBody NicknameRequestDto nicknameRequestDto) {
-        boolean exists = memberService.isNicknameExist(nicknameRequestDto.getNickname());
+        boolean exists = authService.isNicknameExist(nicknameRequestDto.getNickname());
 
         if (exists) {
             throw new RacketPuncherException(INVALID_NICKNAME);
