--- src/main/java/com/example/demo/oauth2/dto/AccessToken.java
@@ -1,16 +1,0 @@
-package com.example.demo.oauth2.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AccessToken {
-    private String access_token;
-    private String token_type;
-    private String refresh_token;
-    private long expires_in;
-    private long refresh_token_expires_in;
-}
