--- src/main/java/com/example/demo/oauth2/dto/ProfileDto.java
@@ -1,14 +1,0 @@
-package com.example.demo.oauth2.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class ProfileDto {
-    private String nickname;
-    private String profileImage;
-}
-
