+++ src/test/java/com/example/demo/matching/controller/MatchingControllerTest.java
@@ -23,9 +23,9 @@ import com.example.demo.matching.dto.MatchingPreviewDto;
 import com.example.demo.matching.service.MatchingService;
 import com.example.demo.openfeign.dto.address.AddressResponseDto;
 import com.example.demo.openfeign.service.address.AddressService;
-import com.example.demo.siteuser.security.JwtAuthenticationFilter;
-import com.example.demo.siteuser.security.SecurityConfiguration;
-import com.example.demo.siteuser.security.TokenProvider;
+import com.example.demo.oauth.security.JwtAuthenticationFilter;
+import com.example.demo.oauth.security.SecurityConfiguration;
+import com.example.demo.oauth.security.TokenProvider;
 import com.example.demo.type.AgeGroup;
 import com.example.demo.type.MatchingType;
 import com.example.demo.type.Ntrp;
